requirements.txt
#devops
black==22.3.0
#click==8.1.3 
pytest==7.4.0
pytest-cov==4.0.0
pylint==2.15.3
requests
#boto3==1.24.87
databricks-sql-connector
#web
#fastapi==0.85.0
#uvicorn==0.18.3
python-dotenv
setuptools
pandas
pyspark
matplotlib
seaborn
ruff
flask